refactor(core,command): move daemon lifecycle out of ipc-client into command (fixes #160)
Split core/src/ipc-client.ts into pure IPC transport (core) and daemon
lifecycle management (command/src/daemon-lifecycle.ts). Extract rawFetch
helper to deduplicate triplicated fetch boilerplate. Consolidate three
overlapping test files into two focused spec files.
